Add GET /api/tweet/:tweet_id logic
I didn't think about use case of going directly to a tweet since
I was focused on tweets_by_handle and tweets_by_email. Of course
it'd be likely someone goes directly to a tweet based on ID, in which
case there would be a respective table/query needed with ID as PK.

I created the route and added the respective test.