Implement basic timeline GET request
Implementation of timeline is very basic right now and doesn't do any
sort of pagination.

In implementation of this I also realized a very basic error with how
I had been setting up my denormalized tweets_by_[XXX]. Updates to user
settings like avatar or alias will be a pain to deal with later on.

I'll need to figure out a way to refactor to accommodate for this, but
the method is slowly approaching a relational mentality and I'm not sure
how far I'm supposed to dig there. I could for instance make a Cassnadra
table for user data, and query it for each tweet. The number of tweeets
fetched at any time is not more than 100, so the performance shouldn't
be bad.